added support for integer arrays.
vectInit now takes an additional argument, which is an enum defined
in vect.h, used to specify the type of the managed array: I_ARRAY (int
array) or C_ARRAY (char array).

Based on this, vectInit will allocate memory for either an integer or
a char array. The array that will be allocated memory has been changed
from a struct member ('dynArray', previously), to a union member.
dynarray is now a union having two members -- i and c -- and it's one of
these that memory will be allocated to, with i being an int* and c
being a char*.

After initialization, there are separate functions that need to be
called to handle a vector managing an inner char array and one managing
an int array, respectively, e.g. c_vectAppend() vs i_vectAppend().