Added automatic memory shrinking capabilities.
The managed array will be reallocated a chunk of memory half what it
currently has when the array is filled to less than half the capacity.

As a way of minimizing code redundancy, I wrote a few functions that
check and resize the array. They're not meant to be called directly,
but are only called by the other functions automatically.
They are private_vectCheckSize_shrink, private_vectCheckSize_grow,
private_vectShrink, private_vectGrow. The first 2 check if the size
of the array is to be halved or doubled, respectively, and the last 2
actually call realloc to halve or double the memory allocated to the
array.

Also added a vectRanRem, which deletes a whole range of values specified
by a starting_index and an ending_index parameter, respectively.